lint fixes
