+++ package.json
@@ -141,7 +141,7 @@
     "react-app-polyfill": "^0.1.3",
     "react-app-rewire-aliases": "^0.2.0",
     "react-app-rewire-hot-loader": "^2.0.1",
-    "react-app-rewired": "^2.1.5",
+    "react-app-rewired": "^2.1.8",
     "react-art": "^16.8.1",
     "react-cropper": "^1.2.0",
     "react-dev-utils": "^11.0.4",
@@ -255,16 +255,16 @@
     "cypress-xpath": "^1.4.0",
     "detox": "^17.6.1",
     "documentation": "9.3.1",
-    "eslint": "^7.14.0",
-    "eslint-config-prettier": "^3.6.0",
-    "eslint-config-react-app": "^3.0.7",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-react-app": "^6.0.0",
     "eslint-plugin-dependencies": "^2.4.0",
-    "eslint-plugin-flowtype": "2.50.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "6.1.2",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "7.11.1",
-    "eslint-plugin-react-hooks": "^1.4.0",
+    "eslint-plugin-flowtype": "^6.1.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.26.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "fake-indexeddb": "^3.1.3",
     "faker": "^4.1.0",
     "fork-ts-checker-webpack-plugin-alt": "^0.4.14",
